add maintenance status note to readme
@chrisbolin we're adding maintenance status to all our OSS. I think `react-fast-compare` should probably be listed as "Active". I also considered "Stable". If you think that's more appropriate, please replace my note with:

```
Maintenance Status: Stable

Formidable is not planning to develop any new features for this project. We are still responding to bug reports and security concerns. We are still welcoming PRs for this project, but PRs that include new features should be small and easy to integrate and should not include breaking changes.
```

Let me know if you have any questions